Added in Nighttime Skybox, Ambient Light, and lighting setting from demo to Environment_Test
I’ve added in a nighttime skybox to the environment_test scene.  I also
added in the nighttime lighting settings from the lighting demo earlier
in the quarter.  There were several directional lights already present
in the scene and that was conflicting with the ambientLight, so I
turned them off.